rename: manage -> manage-list
