+++ src/pages/reservation/manage-detail/index.tsx
@@ -1,0 +1,3 @@
+export default function ManageDetailPage() {
+  return <p>df</p>
+}
